[ MOSIP-26180 ] updated IDP deployment scripts

+++ helm/idp/install.sh
@@ -14,19 +14,26 @@ kubectl create ns $NS
 
 echo Istio label
 kubectl label ns $NS istio-injection=enabled --overwrite
-helm repo update
 
+read -p "Please provide IDP host (ex: idp.sandbox.xyz.net ): " IDP_HOST
+
+if [ -z $IDP_HOST ]; then
+  echo "IDP host not provided; EXITING;"
+  exit 1;
+fi
+
+kubectl -n default get cm global -o json | jq ".data[\"mosip-idp-host\"]=\"$IDP_HOST\"" | kubectl apply -f -
 echo Copy configmaps
 ./copy_cm.sh
+./keycloak-init.sh
 
-echo "Create configmaps idp-ui-cm, delete if exists"
-kubectl -n $NS delete --ignore-not-found=true configmap idp-ui-cm
-kubectl -n $NS create configmap idp-ui-cm --from-literal="REACT_APP_API_BASE_URL=http://$IDP_HOST/v1/idp" --from-literal="REACT_APP_SBI_DOMAIN_URI=http://$IDP_HOST"
 
 echo "Create configmaps mockida, delete if exists"
 kubectl -n $NS  --ignore-not-found=true delete cm mock-auth-data
 kubectl -n $NS create cm mock-auth-data --from-file=./mock-auth-data/
 
+echo Update IDP chart
+helm dependency update
 echo Installing IDP
 helm -n $NS install idp . --version $CHART_VERSION
